python-scripts/requirements.txt
Package    Version
---------- -------
lgpio      0.2.2.0
pip        23.0.1
pyLoRa     0.3.1
rpi-lgpio  0.6
setuptools 66.1.1
spidev     3.6
